Rename package to copyormovecontent
